Upgrade Needle dependency to fix errors
Prior to v0.7.9, Needle depended on a floating version of the "qs" library.
The latest release of "qs" depends on ECMAScript6, which causes failures
when running in any older Node environment.

Upgrading to the latest Needle fixes this issue. There are no compatibility
changes between v0.6.1 and v0.11.0 that should cause issues:
https://github.com/tomas/needle/releases